+++ packages/authentication-local/test/fixture.js
@@ -34,7 +34,16 @@ module.exports = (app = feathers()) => {
     before: {
       create: hashPassword('password')
     },
-    after: protect('password')
+    after: {
+      all: protect('password'),
+      get: [context => {
+        if (context.params.provider) {
+          context.result.fromGet = true;
+        }
+
+        return context;
+      }]
+    }
   });
 
   return app;