+++ packages/authentication/src/jwt.ts
@@ -1,4 +1,5 @@
 import { NotAuthenticated } from '@feathersjs/errors';
+import { omit } from 'lodash';
 import { AuthenticationRequest } from './core';
 import { Params } from '@feathersjs/feathers';
 import { IncomingMessage } from 'http';
@@ -36,14 +37,20 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
    * @param params Service call parameters
    */
   async getEntity (id: string, params: Params) {
+    const { entity } = this.configuration;
     const entityService = this.entityService;
 
     if (entityService === null) {
       throw new NotAuthenticated(`Could not find entity service`);
     }
 
-    // @ts-ignore
-    return entityService.get(id, params);
+    const result = await entityService.get(id, omit(params, 'provider'));
+
+    if (!params.provider) {
+      return result;
+    }
+
+    return entityService.get(id, { ...params, [entity]: result });
   }
 
   async authenticate (authentication: AuthenticationRequest, params: Params) {
@@ -51,7 +58,7 @@ export class JWTStrategy extends AuthenticationBaseStrategy {
     const { entity } = this.configuration;
 
     if (!accessToken) {
-      throw new NotAuthenticated('Not authenticated');
+      throw new NotAuthenticated('No access token');
     }
 
     const payload = await this.authentication.verifyAccessToken(accessToken, params.jwt);